+++ src/frontend/screens/Security/index.tsx
@@ -3,10 +3,8 @@ import {defineMessages, useIntl} from 'react-intl';
 
 import {useSecurityContext} from '../../contexts/SecurityContext';
 import {NativeNavigationComponent} from '../../sharedTypes/navigation';
-import {
-  FullScreenMenuList,
-  MenuListItemType,
-} from '../../sharedComponents/MenuList';
+import {FullScreenMenuList} from '../../sharedComponents/MenuList/FullScreenMenuList';
+import {MenuListItemType} from '../../sharedComponents/MenuList/MenuListItem';
 
 const m = defineMessages({
   title: {
