+++ src/frontend/screens/Settings/ProjectSettings/index.tsx
@@ -1,10 +1,8 @@
 import * as React from 'react';
 import {defineMessages, useIntl} from 'react-intl';
 import {NativeNavigationComponent} from '../../../sharedTypes/navigation';
-import {
-  FullScreenMenuList,
-  MenuListItemType,
-} from '../../../sharedComponents/MenuList';
+import {FullScreenMenuList} from '../../../sharedComponents/MenuList/FullScreenMenuList';
+import {MenuListItemType} from '../../../sharedComponents/MenuList/MenuListItem';
 
 const m = defineMessages({
   title: {
